requirements.txt
flask>=3.0
pandas>=2.0
openpyxl>=3.1
requests>=2.31
python-dotenv>=1.0
